Load recipe's in new thread, add loading indicator [...]
Obviously the sleep will be removed later. While the recipe's could indeed be loaded at mod startup, it's extremely fast so I've not bothered. Additionally, this loader will be used in other cases i.e. change sorting, advanced search, and so on - all of which may take a half second, and a loading indicator is better than a blank or dirty GUI.